Perbaruan chart

+++ midterm_exam_mobile/lib/models/gender.dart
@@ -1,5 +1,5 @@
 class Gender {
   Gender(this.jk, this.jumlah);
   String jk;
-  double jumlah;
+  int jumlah;
 }
